[#2070] Move rating.py into model.
