+++ ckan/model/package.py
@@ -24,6 +24,7 @@ __all__ = ['Package', 'package_table', 'package_revision_table',
            'PACKAGE_NAME_MAX_LENGTH', 'PACKAGE_NAME_MIN_LENGTH',
            'PACKAGE_VERSION_MAX_LENGTH', 'PackageTagRevision', 'PackageRevision']
 
+
 PACKAGE_NAME_MAX_LENGTH = 100
 PACKAGE_NAME_MIN_LENGTH = 2
 PACKAGE_VERSION_MAX_LENGTH = 100
@@ -557,6 +558,55 @@ class Package(vdm.sqlalchemy.RevisionedObjectMixin,
                 context={'model':ckan.model})
         return activity.ActivityDetail(activity_id, self.id, u"Package", activity_type,
             {'package': package_dict })
+
+    def set_rating(self, user_or_ip, rating):
+        '''Record a user's rating of this package.
+
+        The caller function is responsible for doing the commit.
+
+        If a rating is outside the range MAX_RATING - MIN_RATING then a
+        RatingValueException is raised.
+
+        @param user_or_ip - user object or an IP address string
+        '''
+        user = None
+        from user import User
+        from rating import Rating, MAX_RATING, MIN_RATING
+        if isinstance(user_or_ip, User):
+            user = user_or_ip
+            rating_query = meta.Session.query(Rating)\
+                               .filter_by(package=self, user=user)
+        else:
+            ip = user_or_ip
+            rating_query = meta.Session.query(Rating)\
+                               .filter_by(package=self, user_ip_address=ip)
+
+        try:
+            rating = float(rating)
+        except TypeError:
+            raise RatingValueException
+        except ValueError:
+            raise RatingValueException
+        if rating > MAX_RATING or rating < MIN_RATING:
+            raise RatingValueException
+
+        if rating_query.count():
+            rating_obj = rating_query.first()
+            rating_obj.rating = rating
+        elif user:
+            rating = Rating(package=self,
+                            user=user,
+                            rating=rating)
+            meta.Session.add(rating)
+        else:
+            rating = Rating(package=self,
+                            user_ip_address=ip,
+                            rating=rating)
+            meta.Session.add(rating)
+
+
+class RatingValueException(Exception):
+    pass
 
 # import here to prevent circular import
 import tag
