+++ ckan/model/rating.py
@@ -8,7 +8,11 @@ import user
 import domain_object
 import types as _types
 
-__all__ = ['Rating']
+__all__ = ['Rating', 'MIN_RATING', 'MAX_RATING']
+
+MIN_RATING = 1.0
+MAX_RATING = 5.0
+
 
 rating_table = Table('rating', meta.metadata,
                      Column('id', types.UnicodeText, primary_key=True, default=_types.make_uuid),
